backend/tests/test_fastapi_guard_middleware.py
"""
Unit tests for FastAPI-Guard middleware functionality.

Tests verify that the SecurityMiddleware correctly:
1. Allows US IP addresses when publicly deployed
2. Blocks non-US IP addresses when publicly deployed
3. Bypasses geo-blocking when not publicly deployed (dev/local)
4. Handles edge cases and error conditions
"""

import pytest
from unittest.mock import Mock, patch
from fastapi import FastAPI
from fastapi.testclient import TestClient

# Import the guard components - these should be available from fastapi-guard package
try:
    from guard import SecurityMiddleware, SecurityConfig, GeoIPHandler
    GUARD_AVAILABLE = True
except ImportError:
    GUARD_AVAILABLE = False
    # Fallback for testing - create mock classes if guard is not available
    class SecurityMiddleware:
        def __init__(self, app, config):
            self.app = app
            self.config = config

        async def __call__(self, scope, receive, send):
            # Pass through - mock doesn't actually filter
            await self.app(scope, receive, send)

    class SecurityConfig:
        def __init__(self, geo_ip_handler=None, ipinfo_token=None, whitelist_countries=None,
                     trusted_proxies=None, emergency_mode=False, enforce_https=False, passive_mode=False):
            self.geo_ip_handler = geo_ip_handler
            self.ipinfo_token = ipinfo_token
            self.whitelist_countries = whitelist_countries or []
            self.trusted_proxies = trusted_proxies or []
            self.emergency_mode = emergency_mode
            self.enforce_https = enforce_https
            self.passive_mode = passive_mode

    class GeoIPHandler:
        @staticmethod
        def get_country(ip):
            return "US"


class TestFastAPIGuardMiddleware:
    """Test suite for FastAPI-Guard middleware functionality."""

    @pytest.fixture
    def mock_settings_publicly_deployed(self):
        """Mock settings for publicly deployed environment. Should be affected by middleware."""
        mock_settings = Mock()
        mock_settings.is_publicly_deployed = True
        mock_settings.IPINFO_TOKEN = "test_token"
        mock_settings.APP_URL = "https://example.com"
        mock_settings.API_URL = "https://api.example.com"
        mock_settings.COOKIE_SECRET = "test_secret"
        mock_settings.ORIGIN = "example.com"
        return mock_settings

    @pytest.fixture
    def mock_settings_local_dev(self):
        """Mock settings for local development environment. Shouldn't be affected by middleware."""
        mock_settings = Mock()
        mock_settings.is_publicly_deployed = False
        mock_settings.IPINFO_TOKEN = "test_token"
        mock_settings.APP_URL = "http://localhost:3000"
        mock_settings.API_URL = "http://localhost:8000"
        mock_settings.COOKIE_SECRET = "test_secret"
        mock_settings.ORIGIN = "localhost"
        return mock_settings

    @pytest.fixture
    def app_with_guard_middleware_passive(self, mock_settings_publicly_deployed):
        """Create FastAPI app with SecurityMiddleware in passive mode for testing allowed requests as TestClient's IP as "testclient."""
        app = FastAPI()
        
        # Add the SecurityMiddleware with US whitelist in passive mode
        config = SecurityConfig(
            geo_ip_handler=GeoIPHandler,
            ipinfo_token=mock_settings_publicly_deployed.IPINFO_TOKEN,
            whitelist_countries=["US"],
            # Local and Private IPs
            trusted_proxies=["127.0.0.0/8", "::1/128", "10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16"],
            emergency_mode=False,
            enforce_https=False,
            passive_mode=True,  # Passive mode - logs but doesn't block violations
        )
        app.add_middleware(SecurityMiddleware, config=config)
        
        @app.get("/test")
        async def test_endpoint():
            return {"message": "success"}

        return app

    @pytest.fixture
    def app_with_guard_middleware_active(self, mock_settings_publicly_deployed):
        """Create FastAPI app with SecurityMiddleware in active mode for testing blocked requests."""
        app = FastAPI()
        
        # Add the SecurityMiddleware with US whitelist in active mode
        config = SecurityConfig(
            geo_ip_handler=GeoIPHandler,
            ipinfo_token=mock_settings_publicly_deployed.IPINFO_TOKEN,
            whitelist_countries=["US"],
            # Local and Private IPs
            trusted_proxies=["127.0.0.0/8", "::1/128", "10.0.0.0/8", "172.16.0.0/12", "192.168.0.0/16"],
            emergency_mode=False,
            enforce_https=False,
            passive_mode=False,  # Active mode - actually blocks requests
        )
        app.add_middleware(SecurityMiddleware, config=config)
        
        @app.get("/test")
        async def test_endpoint():
            return {"message": "success"}
        
        return app

    @pytest.fixture
    def app_without_guard_middleware(self, mock_settings_local_dev):
        """Create FastAPI app without SecurityMiddleware for local dev testing."""
        app = FastAPI()
        
        @app.get("/test")
        async def test_endpoint():
            return {"message": "success"}
        
        return app

    @pytest.mark.skipif(not GUARD_AVAILABLE, reason="fastapi-guard package not installed")
    @patch('utils.config_utils.get_settings')
    def test_us_ip_allowed_when_publicly_deployed(self, mock_get_settings, app_with_guard_middleware_passive, mock_settings_publicly_deployed):
        """Test that US IP addresses are allowed when publicly deployed."""
        mock_get_settings.return_value = mock_settings_publicly_deployed

        with patch.object(GeoIPHandler, 'get_country', return_value='US'):
            client = TestClient(app_with_guard_middleware_passive)
            response = client.get("/test", headers={"X-Forwarded-For": "8.8.8.8"})

            assert response.status_code == 200
            assert response.json() == {"message": "success"}

    @pytest.mark.skipif(not GUARD_AVAILABLE, reason="fastapi-guard package not installed")
    @patch('utils.config_utils.get_settings')
    @patch('guard.utils.extract_client_ip')
    def test_non_us_ip_blocked_when_publicly_deployed(self, mock_extract_ip, mock_get_settings, app_with_guard_middleware_active, mock_settings_publicly_deployed):
        """Test that non-US IP addresses are blocked when publicly deployed."""
        mock_get_settings.return_value = mock_settings_publicly_deployed
        mock_extract_ip.return_value = "1.2.3.4"  # Mock a non-US IP

        with patch.object(GeoIPHandler, 'get_country', return_value='CA'):
            client = TestClient(app_with_guard_middleware_active)
            response = client.get("/test", headers={"X-Forwarded-For": "1.2.3.4"})

            # Should be blocked by the middleware
            assert response.status_code == 403

    @pytest.mark.skipif(not GUARD_AVAILABLE, reason="fastapi-guard package not installed")
    @patch('utils.config_utils.get_settings')
    @patch('guard.utils.extract_client_ip')
    def test_multiple_non_us_countries_blocked(self, mock_extract_ip, mock_get_settings, app_with_guard_middleware_active, mock_settings_publicly_deployed):
        """Test that various non-US countries are blocked."""
        mock_get_settings.return_value = mock_settings_publicly_deployed

        blocked_countries = ['CA', 'GB', 'DE', 'FR', 'JP', 'CN', 'RU', 'BR']

        for i, country_code in enumerate(blocked_countries):
            mock_extract_ip.return_value = f"1.2.3.{i+10}"  # Different IP for each test
            with patch.object(GeoIPHandler, 'get_country', return_value=country_code):
                client = TestClient(app_with_guard_middleware_active)
                response = client.get("/test", headers={"X-Forwarded-For": f"1.2.3.{i+10}"})

                assert response.status_code == 403, f"Country {country_code} should be blocked"

    @patch('utils.config_utils.get_settings')
    def test_local_dev_bypasses_geo_blocking(self, mock_get_settings, app_without_guard_middleware, mock_settings_local_dev):
        """Test that local development bypasses geo-blocking entirely."""
        mock_get_settings.return_value = mock_settings_local_dev
        
        # Even with a non-US IP, should work in local dev
        client = TestClient(app_without_guard_middleware)
        response = client.get("/test")
        
        assert response.status_code == 200
        assert response.json() == {"message": "success"}

    @pytest.mark.skipif(not GUARD_AVAILABLE, reason="fastapi-guard package not installed")
    @patch('utils.config_utils.get_settings')
    def test_case_insensitive_country_codes(self, mock_get_settings, app_with_guard_middleware_passive, mock_settings_publicly_deployed):
        """Test that country code matching is case insensitive."""
        mock_get_settings.return_value = mock_settings_publicly_deployed

        us_variations = ['US', 'us', 'Us', 'uS']

        for country_code in us_variations:
            with patch.object(GeoIPHandler, 'get_country', return_value=country_code):
                client = TestClient(app_with_guard_middleware_passive)
                response = client.get("/test", headers={"X-Forwarded-For": "8.8.8.8"})

                assert response.status_code == 200, f"Country code {country_code} should be accepted"

    def test_security_config_initialization(self):
        """Test that SecurityConfig initializes correctly with required parameters."""
        config = SecurityConfig(
            geo_ip_handler=GeoIPHandler,
            ipinfo_token="test_token",
            whitelist_countries=["US"]
        )
        
        assert config.geo_ip_handler == GeoIPHandler
        assert config.ipinfo_token == "test_token"
        assert config.whitelist_countries == ["US"]

    def test_security_config_multiple_countries(self):
        """Test SecurityConfig with multiple whitelisted countries."""
        config = SecurityConfig(
            geo_ip_handler=GeoIPHandler,
            ipinfo_token="test_token",
            whitelist_countries=["US", "CA", "GB"]
        )
        
        assert "US" in config.whitelist_countries
        assert "CA" in config.whitelist_countries
        assert "GB" in config.whitelist_countries
        assert len(config.whitelist_countries) == 3

    def test_main_app_middleware_configuration_publicly_deployed(self):
        """Test that main.py correctly adds SecurityMiddleware when publicly deployed."""
        # Mock settings for publicly deployed environment
        mock_settings = Mock()
        mock_settings.is_publicly_deployed = True
        mock_settings.IPINFO_TOKEN = "test_token"
        mock_settings.APP_URL = "https://example.com"
        mock_settings.API_URL = "https://api.example.com"
        mock_settings.COOKIE_SECRET = "test_secret"
        mock_settings.ORIGIN = "example.com"
        
        # Test that the configuration can be created correctly
        config = SecurityConfig(
            geo_ip_handler=GeoIPHandler,
            ipinfo_token="test_token",
            whitelist_countries=["US"]
        )
        
        # Verify the config can be created with expected parameters
        assert config.ipinfo_token == "test_token"
        assert config.whitelist_countries == ["US"]
